Implement JSON-driven theme loader in ThemeService
- Added ThemeService class to manage theme loading and application.
- Implemented methods for scanning built-in and user-defined themes.
- Introduced functionality to validate theme JSON files against a schema.
- Developed color parsing and mapping to legacy ThemeColors structure.
- Enabled theme activation with support for extends chains and token colors.
- Integrated QFileSystemWatcher for dynamic theme updates.
- Added methods for exporting, installing, and uninstalling themes.
- Implemented user and workspace themes directory resolution.

+++ middle/src/pgo/profile_data.cpp
@@ -842,7 +842,8 @@ PGOOptimizer::OptimizationReport PGOOptimizer::GenerateReport() const {
   }
 
   report.estimated_speedup_percent =
-      total_benefit / std::max<double>(1.0, profile_.GetTotalInvocationCount()) * 100.0;
+      total_benefit /
+      std::max<double>(1.0, static_cast<double>(profile_.GetTotalInvocationCount())) * 100.0;
 
   return report;
 }
